move k8s resources into subpackages based on api
