+++ tools/c7n_kube/c7n_kube/resources/apps/statefulset.py
