+++ tools/c7n_kube/c7n_kube/resources/core/namespace.py
