+++ tools/c7n_kube/c7n_kube/resources/core/pod.py
@@ -18,6 +18,7 @@ from c7n_kube.provider import resources
 
 @resources.register('pod')
 class Pod(QueryResourceManager):
+
     class resource_type(TypeInfo):
         group = 'Core'
         version = 'V1'
